Fix update when not urls

+++ UpdateTimeTable.py
@@ -21,26 +21,27 @@ try:
     # скачиваем excel
     html = hse.get_html('https://perm.hse.ru/forstudents/timetable/')
     urls = hse.get_urls(html)
-    data_pars = hse.download_files(urls)
+    if len(urls) != 0:
+        data_pars = hse.download_files(urls)
 
-    # достаем даты занятий
-    date = []
-    groups = []
-    for group_pars in data_pars:
-        for pars in group_pars:
-            for par in pars:
-                groups.append(par.group)
-                date.append(par.date_lesson)
-    # удаляем из БД данные по текущим датам
-    date = list(set(date))
-    groups = list(set(groups))
-    db.delete_timetables_date(date, groups)
+        # достаем даты занятий
+        date = []
+        groups = []
+        for group_pars in data_pars:
+            for pars in group_pars:
+                for par in pars:
+                    groups.append(par.group)
+                    date.append(par.date_lesson)
+        # удаляем из БД данные по текущим датам
+        date = list(set(date))
+        groups = list(set(groups))
+        db.delete_timetables_date(date, groups)
 
-    # Заполняем новыми данными
-    for group_pars in data_pars:
-        for pars in group_pars:
-            for par in pars:
-                db.insert_timetables(par)
-    # tg.send_msg(admin_id, 'TimeTables is update success')
+        # Заполняем новыми данными
+        for group_pars in data_pars:
+            for pars in group_pars:
+                for par in pars:
+                    db.insert_timetables(par)
+        # tg.send_msg(admin_id, 'TimeTables is update success')
 except Exception as error:
     tg.send_msg(admin_id, 'ERROR:\n' + str({error}))
